# goswagger/swagger.yaml
openapi: 3.0.3
info:
  title: Karygo APIs Documentation
  description: |-
    Current consumable APIs for Karygo developers

    Some useful tips:
    - Ensure you change the Base Url to [localhost](http://localhost:8000) for local testing
    - Documentation is done in YAML file
    - The Base Url used for this documentation is [here](https://karygo.herokuapp.com)
  version: 1.0.0
servers:
  - url: https://karygo.herokuapp.com
  - url: http://localhost:8000
...
paths:
...
  /admin/{adminId}:
    put:
      tags:
        - Admin
      summary: Update an existing Admin
      description: Should ONLY be accessible when a admin has logged in
      operationId: updateAdmin
      parameters:
        - name: adminId
          in: path
          description: userId to be updated
          required: true
          explode: true
          schema:
            type: integer

      requestBody:
        description: Update an existing admin in Admin table
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Admin"
          application/xml:
            schema:
              $ref: "#/components/schemas/Admin"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
            application/xml:
              schema:
                $ref: "#/components/schemas/Admin"
        "400":
          description: Invalid ID supplied
        "404":
          description: page not found
        "405":
          description: Validation exception
      security:
        - karygo_auth:
            - write:sample
            - read:sample
    get:
      tags:
        - Admin
      summary: Get an existing admin
      description: Get admin by Id
      operationId: getAdmin
      parameters:
        - name: adminId
          in: path
          description: adminId to be fetched
          required: true
          explode: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
            application/xml:
              schema:
                $ref: "#/components/schemas/Admin"
        "400":
          description: Invalid ID supplied
        "404":
          description: page not found
        "405":
          description: Validation exception
      security:
        - karygo_auth:
            - write:sample
            - read:sample
    delete:
      tags:
        - Admin
      summary: Deletes an admin
      description: delete an admin by Id
      operationId: deleteAdmin
      parameters:
        - name: api_key
          in: header
          description: ""
          required: false
          schema:
            type: string
        - name: adminId
          in: path
          description: admin ID to be deleted
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Operation successful
        "400":
          description: Invalid adminId
      security:
        - karygo_auth:
            - write:userR
            - read:userW
  /admin:
    post:
      tags:
        - Admin
      summary: Add a new admin to the Admin table
      description: Add a new admin
      operationId: createAdmin
      requestBody:
        description: Create a new admin in Admin table
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Admin"
          application/xml:
            schema:
              $ref: "#/components/schemas/Admin"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Admin"
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
            application/xml:
              schema:
                $ref: "#/components/schemas/Admin"
        "405":
          description: Invalid input
      security:
        - karygo_auth:
            - write:sample
            - read:sample
    get:
      tags:
        - Admin
      summary: Get all admins from Admin table
      description: Get all Admin
      operationId: allAdmin
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Admin"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Admin"
        "404":
          description: Page not found
      security:
        - karygo_auth:
            - write:contactsW
            - read:contactsR
...
  /admin/transaction/{txId}:
    get:
      tags:
        - Admin
      summary: Get an existing transaction
      description: Get transaction by Id
      operationId: getTransaction
      parameters:
        - name: txId
          in: path
          description: transactionId to be fetched
          required: true
          explode: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
            application/xml:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Invalid ID supplied
        "404":
          description: page not found
        "405":
          description: Validation exception
      security:
        - karygo_auth:
            - write:sample
            - read:sample
components:
  schemas:
    Transaction:
      type: object
      required:
        - id
        - travelerId
        - senderId
        - itemId
        - tripId
      properties:
        ID:
          type: integer
          format: int64
          example: 10
        CreatedAt:
          type: string
          example: "0001-01-01T00:00:00Z"
        UpdatedAt:
          type: string
          example: "0001-01-01T00:00:00Z"
        DeletedAt:
          type: string
          example: "0001-01-01T00:00:00Z"
        travelerId:
          type: integer
          format: int64
          example: 198772
        senderId:
          type: integer
          format: int64
          example: 57789
        itemId:
          type: integer
          format: int64
        tripId:
          type: integer
          example: 34556
        successCode:
          type: string
          example: "YT9965458YZ"
        status:
          type: string
          example: completed
          enum:
            - completed
            - pending
            - canceled
        complete:
          type: boolean
      xml:
        name: transaction
    Lga:
      type: object
      properties:
        ID:
          type: integer
          format: int64
          example: 100000
        CreatedAt:
          type: string
          example: "0001-01-01T00:00:00Z"
        UpdatedAt:
          type: string
          example: "0001-01-01T00:00:00Z"
        DeletedAt:
          type: string
          example: "0001-01-01T00:00:00Z"
        name:
          type: string
          example: "Eti-Osa"
        stateId:
          type: integer
          example: 10
        state:
          type: object
          items:
            $ref: "#/components/schemas/State"
      xml:
        name: lga
    State:
      type: object
      properties:
        id:
          type: integer
          example: 10
        name:
          type: string
          example: "Lagos"
      xml:
        name: state
...
    Admin:
      type: object
      properties:
        ID:
          type: integer
          format: int64
        CreatedAt:
          type: string
          example: "0001-01-01T00:00:00Z"
        UpdatedAt:
          type: string
          example: "0001-01-01T00:00:00Z"
        DeletedAt:
          type: string
          example: "0001-01-01T00:00:00Z"
        firstName:
          type: string
          example: "john"
        lastName:
          type: string
          example: "doe"
        email:
          type: string
          example: "johndoe@gmail.com"
        phone:
          type: string
          example: "2340033335566"
        adminType:
          type: string
          example: moderator
          enum:
            - superUser
            - moderator
            - user
        jobTitle:
          type: string
        department:
          type: string
    User:
      type: object
      properties:
        ID:
          type: integer
          format: int64
          example: 10
        CreatedAt:
          type: string
          example: "0001-01-01T00:00:00Z"
        UpdatedAt:
          type: string
          example: "0001-01-01T00:00:00Z"
        DeletedAt:
          type: string
          example: "0001-01-01T00:00:00Z"
        firstName:
          type: string
          example: "John"
        middeName:
          type: string
          example: "doe"
        lastName:
          type: string
          example: "James"
        email:
          type: string
          example: johndoe@email.com
        password:
          type: string
          example: "12345"
        address:
          type: string
          example: "21st street, NY, USA."
        gender:
          type: string
          example: "Male"
        lgaId:
          type: integer
          format: int64
          example: 10
        lga:
          type: object
          items:
            $ref: "#/components/schemas/Lga"
      xml:
        name: user
...
  securitySchemes:
    karygo_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://karygo.sample.io/oauth/authorize
          scopes:
            write: modify data in the table
            read: read data in the table
    api_key:
      type: apiKey
      name: api_key
      in: header